Change NInt to TInt; change hierarchy
Replaces the NInt kind with TInt, which can represent either a positive
or a negative integer.  This will be used for the result kind of
addition and subtraction of a negative integer and a positive integer so
that type checking can succeed.

TypeInts and TypeRats are now subordinate to TypeNums in the hierarchy
for greater clarity when looking at the haddocks - TypeNums is the main
parent module to import unless you explicitly want to import just one of
them.